Move benchmarks to examples
