+++ examples/experimental/torch/mmllms_capability_eval/mmebench.py
@@ -25,7 +25,6 @@ from transformers import AutoProcessor
 from transformers import set_seed
 
 from nncf.experimental.torch.genai_optimizations import get_inputs_embeds
-from nncf.experimental.torch.genai_optimizations.benchmarks.utils import add_visual_pruning_args
 
 
 class MetricCalculator:
@@ -84,28 +83,6 @@ class MetricCalculator:
             "other_num": other_num,
             "acc": acc,
         }
-
-
-def get_model_class(model_name):
-    if "Qwen2.5-VL" in model_name:
-        from transformers import Qwen2_5_VLForConditionalGeneration
-
-        return Qwen2_5_VLForConditionalGeneration
-    elif "Qwen2-VL" in model_name:
-        from transformers import Qwen2VLForConditionalGeneration
-
-        return Qwen2VLForConditionalGeneration
-    elif "llava-1.5" in model_name:
-        from transformers import LlavaForConditionalGeneration
-
-        return LlavaForConditionalGeneration
-    elif "llava-v1.6" in model_name:
-        from transformers import LlavaNextForConditionalGeneration
-
-        return LlavaNextForConditionalGeneration
-    else:
-        error_msg = f"Unsupported model class for: {model_name}"
-        raise ValueError(error_msg)
 
 
 def evaluate(args):
@@ -197,6 +174,36 @@ def evaluate(args):
     print(f"\n MME Evaluation for '{category}'")
     for k, v in metrics.items():
         print(f"{k:>12}: {v:.4f}" if isinstance(v, float) else f"{k:>12}: {v}")
+
+
+def get_model_class(model_name):
+    if "Qwen2.5-VL" in model_name:
+        from transformers import Qwen2_5_VLForConditionalGeneration
+
+        return Qwen2_5_VLForConditionalGeneration
+    elif "Qwen2-VL" in model_name:
+        from transformers import Qwen2VLForConditionalGeneration
+
+        return Qwen2VLForConditionalGeneration
+    elif "llava-1.5" in model_name:
+        from transformers import LlavaForConditionalGeneration
+
+        return LlavaForConditionalGeneration
+    elif "llava-v1.6" in model_name:
+        from transformers import LlavaNextForConditionalGeneration
+
+        return LlavaNextForConditionalGeneration
+    else:
+        error_msg = f"Unsupported model class for: {model_name}"
+        raise ValueError(error_msg)
+
+
+def add_visual_pruning_args(parser):
+    group = parser.add_argument_group("Visual Token Pruning Arguments")
+    group.add_argument("--enable_visual_pruning", action="store_true", help="Enable visual token pruning")
+    group.add_argument("--num_keep_tokens", type=int, default=128, help="Number of visual tokens to keep")
+    group.add_argument("--theta", type=float, default=0.5, help="Balance factor for diversity vs relevance")
+    return parser
 
 
 if __name__ == "__main__":
