+++ src/nncf/experimental/torch/genai_optimizations/requirements.txt
@@ -1,6 +1,4 @@
 accelerate==1.9.0
-datasets==2.14.7
 transformers>=4.48.0
 torch==2.8.0
 torchvision==0.23.0
-rouge==1.0.1
